make compileBddForS3 aws generic
