+++ .changes/next-release/feature-e8b0c9d42596402804ac70a348da0fd38622a0e4.json
@@ -1,6 +1,6 @@
 {
   "type": "feature",
-  "description": "Add new transform compileBdd and compileBddForS3",
+  "description": "Add new transform compileBdd and compileBddForAws",
   "pull_requests": [
     "[#2953](https://github.com/smithy-lang/smithy/pull/2953)"
   ]
