+++ smithy-aws-endpoints/src/it/java/software/amazon/smithy/rulesengine/aws/language/functions/S3BddTest.java
@@ -21,8 +21,8 @@ import software.amazon.smithy.model.shapes.ServiceShape;
 import software.amazon.smithy.model.shapes.Shape;
 import software.amazon.smithy.model.shapes.ShapeId;
 import software.amazon.smithy.rulesengine.analysis.BddCoverageChecker;
-import software.amazon.smithy.rulesengine.aws.s3.CompileBddForS3;
 import software.amazon.smithy.rulesengine.aws.s3.S3TreeRewriter;
+import software.amazon.smithy.rulesengine.aws.transforms.CompileBddForAws;
 import software.amazon.smithy.rulesengine.language.EndpointRuleSet;
 import software.amazon.smithy.rulesengine.language.evaluation.TestEvaluator;
 import software.amazon.smithy.rulesengine.logic.bdd.CostOptimization;
@@ -165,7 +165,7 @@ class S3BddTest {
         TransformContext context = TransformContext.builder()
                 .model(model)
                 .build();
-        Model result = new CompileBddForS3().transform(context);
+        Model result = new CompileBddForAws().transform(context);
         Shape serviceShape = result.expectShape(S3_SERVICE_ID);
         assertTrue(serviceShape.hasTrait(EndpointBddTrait.ID));
     }
