+++ smithy-aws-endpoints/src/main/resources/META-INF/services/software.amazon.smithy.build.ProjectionTransformer
@@ -1,1 +1,1 @@
-software.amazon.smithy.rulesengine.aws.s3.CompileBddForS3
+software.amazon.smithy.rulesengine.aws.transforms.CompileBddForAws
