+++ smithy-rules-engine/src/main/resources/META-INF/services/software.amazon.smithy.build.ProjectionTransformer
